Fix 'invisible fill' bug
The drawing issue (invisible fill animation is somehow caused by having blending enabled.
Erased segments being detected as boundaries was due to the (erased) color having an alpha of 0 but not a color value of 0. We can't change the color towards black, so a new branch is added in the fill routine